app/requirements.txt
# Web 框架
fastapi==0.109.0
uvicorn[standard]==0.27.0

# 数据库
sqlalchemy==2.0.25
pymysql==1.1.0

# 认证与安全
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# 其他
cryptography==42.0.0
